model/model1/functions/featuer_make/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-storage
google-cloud-bigquery